clothes_service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
django-filter>=24.0,<25
psycopg2-binary>=2.9,<3
gunicorn>=22.0,<23
